- use int as external voice id

+++ include/YoAudio.h
@@ -11,9 +11,9 @@ bool YOA_API YOA_Init(void);
 
 void YOA_API YOA_Quit(const bool quitSDL);
 
-uint16_t YOA_API YOA_PlayWavFile(const char * filename, const bool loop, const float volume, const float pitch, const float fadeIn);
+int YOA_API YOA_PlayWavFile(const char * filename, const bool loop, const float volume, const float pitch, const float fadeIn);
 
-bool YOA_API YOA_StopVoice(const uint16_t id, const float fadeOut);
+bool YOA_API YOA_StopVoice(const int id, const float fadeOut);
 
 void YOA_API YOA_Pause(void);
